chore(desktop): zero eslint warnings on merged redesign
