+++ crates/visio-desktop/frontend/src/components/ui/useConfirm.ts
@@ -1,0 +1,21 @@
+import { createContext, useContext } from 'react'
+
+export interface ConfirmOptions {
+  title?: string
+  message: string
+  confirmLabel: string
+  cancelLabel: string
+  danger?: boolean
+}
+
+export type ConfirmFn = (opts: ConfirmOptions) => Promise<boolean>
+
+export const ConfirmContext = createContext<ConfirmFn | null>(null)
+
+export function useConfirm(): ConfirmFn {
+  const ctx = useContext(ConfirmContext)
+  if (!ctx) {
+    throw new Error('useConfirm must be used inside <ConfirmProvider>')
+  }
+  return ctx
+}
